fix: merge conflict error
